[WIP] Multitracking

+++ student/trackmanagement.py
@@ -117,14 +117,16 @@ class Trackmanagement:
             if meas_list:  # if not empty
                 if meas_list[0].sensor.in_fov(track.x):
                     # your code goes here
-                    track.score -= 1.0/params.window
+                    track.score -= 1./params.window
+                    print("Track score: {}".format(track.score))
 
         # delete old tracks
 
-        for track in self.track_list:
+        for index in unassigned_tracks:
+            track = self.track_list[index]
             if (track.state == 'confirmed' and track.score <= params.delete_threshold) or \
-                    (track.P[0, 0] > params.max_P) or \
-                    (track.P[1, 1] > params.max_P):
+                    (track.state != 'confirmed' and track.P[0, 0] >= params.max_P) or \
+                    (track.state != 'confirmed' and track.P[1, 1] >= params.max_P):
                 self.delete_track(track)
 
         ############
@@ -155,8 +157,12 @@ class Trackmanagement:
         # - increase track score
         # - set track state to 'tentative' or 'confirmed'
         ############
-        track.score = track.score + (1.0/params.window) if track.state == 'tentative' else track.score
-        track.state = 'confirmed' if track.score >= params.confirmed_threshold else 'tentative'
+        track.score += (1.0/params.window)
+
+        if track.score >= params.confirmed_threshold:
+            track.state = 'confirmed'
+        else:
+            track.state = 'tentative'
         ############
         # END student code
         ############
